Fix an edge case with dynamic indexing
If:
- a document addition or update encounters another new
word that occurs lexicographically before an existing term
- and that existing term's postings list has other terms in
the posting list yet to be encountered in lexicographical order

Then there is a possibility that start offset of the second
existing term would be wrong when the postings list
is processed again as it has already been flushed.

Let's persist the postings list writers.