Apply BM25 to child expressions immediately
BM25 scoring is deferred until the ranking stage,
where it is only computed on the top-level expressions.

This results in the relative ordering of disjunctive expressions
within not being preserved.
E.g. "(foo bar)" is scored by computing BM25 on the merged
term frequencies and idf of the parentheses expression.

This commit changes this to sum the scores of "foo" and "bar"
individually, preserving the relative ordering of documents
in child experssions.