Fix a rare benign race with query freeing
Occurs when a subsequent N request is made to the
worker, but the query is free-ed before the request is
complete. (the worker reply is received)

This is extremely rare as the subsequent N request
is very fast, and does not make any network requests.

The error is benign as it occurs as the last step in
the worker onmessage handler.