Bundle search ui assets into cli tool
The indexer cli tool now outputs search UI
assets, this allows hosting compatible versions
of the search UI and indexer easily.

The mdBook plugin is also refactored to
make use of this, using morsels_indexer as
a library.

This also removes the dependency on having
the indexer installed on the system.

In so doing, 2 obscure options in the mdBook plugin
also had to be removed. (auto-scaling / no-css)