Implement enum storage for fields
It is possible but inefficient to use Morsels' query syntax for
filtering results per a select set of terms, as such terms are
stored using an inverted index.

For finite values unique to each document, these values can
be more efficiently mapped to integers and bitpacked for
compression. A 50000 document collection with 16 such values
for example would only take 25KB to store monolithically.

In this, UI support for multi-select dropdowns is also added.
A massive rewrite of the internal UI code was done to
prevent re-rendering of the multi-select component, along
with the header and other loading states.

The headerRender and loadingIndicatorRender APIs are
also removed here, their utility is likely not much, and
maintaining them is difficult. With the search API, the
point of supporting them should be more moot as well.

An accompanying minor breaking change is the renaming
of the do_store option to storage: ["text"], combining
it with the new "enum" interface.

As part of the UI internal redesign, the tip can now also
be used on mode='target' with proper positioning.